requirements.txt
fastapi
uvicorn
httpx
jinja2
python-dotenv
requests